Add logging infrastructure
Currently, OSD relies on a mixture of different printf()s and assert()s
to log data of various severity levels and report them to the user.
assert()s are bad in a library in most cases, as the application has no
way to recover from a failure inside the library. They should be avoided
therefore, unless very good reasons exists (e.g. OOM asserts).
printfs(), on the other hand, do not enable configurable log levels and
cannot be easily routed to the application which uses the OSD library.

This commits takes large parts of the log infrastructure of
glip/liboptimsochost, which has proven itself over time to work well.